packages/compute/pyproject.toml
[project]
name = "insight-mirror-compute"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "akshare>=1.16.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pgvector>=0.3.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "scipy>=1.12.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
